Extract pages into enum and change types file structure
