+++ client/src/App.tsx
@@ -19,6 +19,7 @@ import Quizzes from './components/quizzes/Quizzes';
 import { AuthProvider } from './contexts/AuthContext';
 import { ErrorProvider } from './contexts/ErrorContext';
 import './styles/style.scss';
+import { PagesEnum } from './types/enums/PagesEnum';
 
 function App() {
   return (
@@ -33,21 +34,21 @@ function App() {
                   <Header />
                   <Routes>
                     {/* Everyone */}
-                    <Route path="/" element={<Home />} />
-                    <Route path="/courses" element={<Courses />} />
-                    <Route path="/blog" element={<Blog />} />
-                    <Route path="/contact" element={<ContactUs />} />
+                    <Route path={PagesEnum.Home} element={<Home />} />
+                    <Route path={PagesEnum.Courses} element={<Courses />} />
+                    <Route path={PagesEnum.Blog} element={<Blog />} />
+                    <Route path={PagesEnum.Contact} element={<ContactUs />} />
 
                     {/* Only guests */}
                     <Route element={<ProtectedRoute onlyUser={false} />}>
-                      <Route path="/login" element={<Login />} />
-                      <Route path="/register" element={<Register />} />
+                      <Route path={PagesEnum.Login} element={<Login />} />
+                      <Route path={PagesEnum.Register} element={<Register />} />
                     </Route>
 
                     {/* Only logged users */}
                     <Route element={<ProtectedRoute onlyUser={true} />}>
-                      <Route path="/logout" element={<Logout />} />
-                      <Route path="/quizzes" element={<Quizzes />} />
+                      <Route path={PagesEnum.Logout} element={<Logout />} />
+                      <Route path={PagesEnum.Quizzes} element={<Quizzes />} />
                     </Route>
 
                     <Route path="*" element={<NotFound />} />
