+++ client/src/components/auth/authenticate/Authenticate.tsx
@@ -1,7 +1,7 @@
 import { PropsWithChildren, useEffect } from 'react';
 import { CachePolicies, useFetch } from 'use-http';
 import { useAuthContext } from '../../../contexts/AuthContext';
-import { IAuthResponse } from '../../../interfaces/IAuthResponse';
+import { IAuthResponse } from '../../../types/interfaces/IAuthResponse';
 import { initialAuthUtils } from '../../../utils/initialAuthUtils';
 import Spinner from '../../common/spinner/Spinner';
