+++ client/src/contexts/AuthContext.tsx
@@ -1,8 +1,8 @@
 import { PropsWithChildren, createContext, useContext } from 'react';
 import { useLocalStorage } from '../hooks/useLocalStorage';
-import { IAuthResponse } from '../interfaces/IAuthResponse';
-import { IAuthStorage } from '../interfaces/IAuthStorage';
-import { IUser } from '../interfaces/IUser';
+import { IAuthResponse } from '../types/interfaces/IAuthResponse';
+import { IAuthStorage } from '../types/interfaces/IAuthStorage';
+import { IUser } from '../types/interfaces/IUser';
 import { isJwtExpired } from '../utils/jwtUtils';
 
 type AuthContextType = {
