+++ client/src/types/enums/PagesEnum.ts
@@ -1,0 +1,10 @@
+export enum PagesEnum {
+  Home = '/',
+  Contact = '/contact',
+  Login = '/login',
+  Register = '/register',
+  Logout = '/logout',
+  Quizzes = '/quizzes',
+  Courses = '/courses',
+  Blog = '/blog',
+}
