+++ client/src/types/interfaces/IAuthResponse.ts
