+++ client/src/types/interfaces/IAuthStorage.ts
