+++ client/src/types/interfaces/IError.ts
