+++ client/src/types/interfaces/IUser.ts
